2019.9.20 02:51
优化了一些实体类的构造方法
把一些字段从int改成了byte
增加了一些测试数据
前端和后端共同完成了一个业务逻辑，任重道远
踩坑webview文字渲染，透明度低于0.9突然变淡，粗体有问题，css穿透有问题

+++ server/src/main/java/model/Diagnosis.java
@@ -34,7 +34,7 @@ public class Diagnosis {
     /*最终诊断信息*/
     private String judgement;
 
-    private int status;
+    private byte status;
 
 
     /**
@@ -50,7 +50,7 @@ public class Diagnosis {
      * @param status         诊断状态
      * @return 一个id自动生成的诊断信息对象
      */
-    public static Diagnosis insert(long registrationID, String complaint, int clazz, String disease, String suggestion, String note, String judgement, int status) {
+    public static Diagnosis insert(long registrationID, String complaint, int clazz, String disease, String suggestion, String note, String judgement, byte status) {
         return new Diagnosis(IDGenerator.generate(), registrationID, complaint, clazz, disease, suggestion, note, judgement, status);
     }
